requirements.txt
beautifulsoup4==4.12.2
begins==0.9
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
ordereddict==1.1
packaging==23.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.7
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-file==1.5.1
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
text-unidecode==1.3
thinc==8.1.10
tinysegmenter==0.3
tldextract==3.4.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
wasabi==1.1.1
Werkzeug==2.3.4
zipp==3.15.0
